Add 'show' button for each country listing
A 'show' button has been added to the display of each country in the list. The onClick handler for this button sets the current search query to the name of the selected country. This functionality gives users the ability to directly view the selected country's details upon clicking the 'show' button.