Add weather data to countries
The commit introduces a new feature in the application where users can view weather data for a selected country. It added a new effect hook dependent on the "countrySelected" variable. When a country is selected, the application fetches and displays the weather data via the "getWeatherForCountry" function. The name of the country selected is also set whenever the "displayCountries" function is invoked.